+++ strategy_pattern/strategy_pattern/IProduct.cs
@@ -5,7 +5,7 @@ namespace Strategy
 {
     public interface IProduct
     {
-        double SellingPrice();
+        Decimal SellingPrice();
 
         bool OnSale();
     }
